requirements.txt
Django==4.2.1
django-cors-headers==4.0.0
djangorestframework==3.14.0
drf-yasg==1.21.5
psycopg2-binary==2.9.6
python-decouple==3.8
